% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfun.R
\name{dfun}
\alias{dfun}
\alias{cdfun}
\alias{autoplot.distribution}
\alias{chart.distribution}
\alias{geom_funfill}
\title{Create and plot density functions for distribution objects}
\usage{
dfun(object, i = 1)

cdfun(object, i = 1)

\method{autoplot}{distribution}(
  object,
  n = 500,
  xlim = NULL,
  size = 99.5,
  xlab = "Quantile",
  ylab = if (type == "density") "Probability density" else
    "Cumulative probability density",
  plot.it = TRUE,
  use.chart = FALSE,
  ...,
  type = "density",
  theme = NULL
)

\method{chart}{distribution}(data, ..., type = "density", env = parent.frame())

geom_funfill(
  mapping = NULL,
  data = NULL,
  fun,
  from,
  to,
  geom = "area",
  fill = "salmon",
  alpha = 0.5,
  ...
)
}
\arguments{
\item{object}{A \strong{distribution} object, as from the \{distributional\}
package.}

\item{i}{The distribution to use from the list (first one by default)}

\item{n}{The number of points to use to draw the density functions (500 by
default) of continuous distributions.}

\item{xlim}{Two numbers that limit the X axis.}

\item{size}{If \verb{xlim=} is not provided, it is automatically calculated using
the size of the CI between 0 and 100 (99.5 by default) for continuous
distributions.}

\item{xlab}{The label of the X axis ("Quantile" by default).}

\item{ylab}{The label of the Y axis ("Probability density" or "Cumulative
probability density" by default).}

\item{plot.it}{Should the densities be plotted for all the distributions
(\code{TRUE} by default)?}

\item{use.chart}{Should \code{\link[=chart]{chart()}} be used (\code{TRUE} by default)? Otherwise,
\code{\link[=ggplot]{ggplot()}} is used.}

\item{...}{Further arguments to \code{\link[=stat_function]{stat_function()}}.}

\item{type}{The type of plot ("density" by default, or "cumulative").}

\item{theme}{The theme for the plot (ignored for now).}

\item{data}{The data frame to use (\code{NULL} by default).}

\item{env}{The environment to use to evaluate expressions.}

\item{mapping}{the mapping to use (\code{NULL} by default.}

\item{fun}{The function to use (could be \code{dfun(distribution_object)}).}

\item{from}{The first quantile to delimit the filled area.}

\item{to}{The second quantile to delimit the filled area.}

\item{geom}{The geom to use (\code{"area"} by default).}

\item{fill}{The color to fill the area (\code{"salmon"} by default).}

\item{alpha}{The alpha transparency to apply, 0.5 by default.}
}
\value{
Either a function or a ggplot object.
}
\description{
The \strong{distribution} objects represent one or more statistical distributions.
The functions \code{\link[=dfun]{dfun()}} and \code{\link[=geom_funfill]{geom_funfill()}}, together with \code{\link[=chart]{chart()}} allow to
plot them.
}
\examples{
library(distributional)
library(chart)
di1 <- dist_normal(mu = 1, sigma = 1.5)
chart(di1) +
  geom_funfill(fun = dfun(di1), from = -5, to = 1)

# With two distributions
di2 <- c(dist_normal(10, 1), dist_student_t(df = 3, 13, 1))
chart(di2) +
  geom_funfill(fun = dfun(di2, 1), from = -5, to = 0) +
  geom_funfill(fun = dfun(di2, 2), from = 2, to = 6, fill = "turquoise3")
chart$cumulative(di2)
# A discrete distribution
di3 <- dist_binomial(size = 7, prob = 0.5)
chart(di3)
chart$cumulative(di3)
# A continuous together with a discrete distribution
di4 <- c(dist_normal(mu = 4, sigma = 2), dist_binomial(size = 8, prob = 0.5))
chart(di4)
chart$cumulative(di4)
}
