% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferit-package.R
\docType{package}
\name{inferit-package}
\alias{inferit}
\alias{inferit-package}
\title{inferit: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org}

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be} (\href{https://orcid.org/0000-0001-9514-1014}{ORCID})
}

}
\keyword{internal}
