% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stddev.R
\name{stddev}
\alias{stddev}
\alias{stddev.default}
\alias{stddev.distribution}
\title{Get standard deviation for a distribution objects}
\usage{
stddev(x, ...)

\method{stddev}{default}(x, ...)

\method{stddev}{distribution}(x, ...)
}
\arguments{
\item{x}{A \strong{distribution} object, as from the \{distributional\} package.}

\item{...}{Further arguments (not used yet).}
}
\value{
A numeric vector with one or more standard deviations.
}
\description{
The \strong{distribution} objects represent one or more statistical distributions.
The generic functions \code{\link[=stddev]{stddev()}} returns the standard deviation for these
distributions.
}
\examples{
library(distributional)
n1 <- dist_normal(mu = 1, sigma = 1.5)
n1
class(n1)
family(n1)
mean(n1)
variance(n1)
stddev(n1)
}
