% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.htest.R
\name{tabularise_default.htest}
\alias{tabularise_default.htest}
\title{Create a ric-formatted table from an htest object}
\usage{
\method{tabularise_default}{htest}(
  data,
  header = TRUE,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An \strong{htest} object}

\item{header}{If \code{TRUE} (by default), add a title to the table}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{show.signif.stars}{If \code{TRUE} (by default), add the significance stars
to the table.}

\item{...}{Additional arguments (unused for now).}

\item{env}{The environment where to evaluate lazyeval expressions (unused for
now).}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Default type to \code{\link[=tabularise]{tabularise()}} an \strong{htest} object as a flextable.
}
\examples{
data(iris)
cor <- cor.test(iris$Sepal.Length, iris$Sepal.Width)
tabularise::tabularise(cor)

tabularise::tabularise(t.test(x = 1:10, y = 7:20), lang = "fr")
}
