% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferit-package.R
\docType{package}
\name{inferit-package}
\alias{inferit}
\alias{inferit-package}
\title{'SciViews::R' - Hypothesis Tests and Statistical Distributions}
\description{
Statistical distributions and hypothesis tests objects with rich-formatted
charts and tables.
}
\section{Important functions}{

\itemize{
\item \code{\link[=tabularise]{tabularise()}} methods for \strong{htest} objects.
\item \code{\link[=stddev]{stddev()}} for \strong{distribution} objects.
\item \code{\link[=dfun]{dfun()}}, \code{\link[=cdfun]{cdfun()}} density function for \strong{distribution} objects.
\item \code{\link[=chart]{chart()}} method for \strong{distribution} objects.
\item \code{\link[=geom_funfill]{geom_funfill()}} fills a part of a distribution density function.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/inferit}
  \item \url{https://www.sciviews.org/inferit/}
  \item Report bugs at \url{https://github.com/SciViews/inferit/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be} (\href{https://orcid.org/0000-0001-9514-1014}{ORCID})
}

}
