% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferit-package.R
\docType{package}
\name{inferit-package}
\alias{inferit-package}
\title{Hypothesis Tests and Statistical Distributions for 'SciViews::R'}
\description{
Statistical distributions and hypothesis tests objects with rich-formatted
charts and tables.
}
\section{Important functions}{

\itemize{
\item \code{\link[=tabularise]{tabularise()}} methods for \strong{htest} objects.
\item \code{\link[=stddev]{stddev()}} for \strong{distribution} objects.
\item \code{\link[=dfun]{dfun()}}, \code{\link[=cdfun]{cdfun()}} density function for \strong{distribution} objects.
\item \code{\link[=chart]{chart()}} method for \strong{distribution} objects.
\item \code{\link[=geom_funfill]{geom_funfill()}} fills a part of a distribution density function.
}
}

