% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.htest.R
\name{tabularise_default.htest}
\alias{tabularise_default.htest}
\title{Method to transform htest object into a flextable object}
\usage{
\method{tabularise_default}{htest}(
  data,
  header = TRUE,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An htest object}

\item{header}{If TRUE add a title to the table}

\item{lang}{The language to use. The default value can be set with, e.g., options(data.io_lang = "fr") for French.}

\item{show.signif.stars}{If TRUE add the sifnificance stars to the table}

\item{...}{Unused argument}

\item{env}{The environment where to evaluate the function}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
This method tabularises an htest object as a flextable.
}
\examples{
library(tabularise)

data(iris)
cor <- cor.test(iris$Sepal.Length, iris$Sepal.Width)
tabularise(cor)

tabularise(t.test(1:10, y = c(7:20)))

}
