% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.htest.R
\name{tabularise_default.htest}
\alias{tabularise_default.htest}
\title{Create a rich-formatted table from an htest object}
\usage{
\method{tabularise_default}{htest}(
  data,
  header = TRUE,
  title = NULL,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An \strong{htest} object}

\item{header}{If \code{TRUE} (by default), add a header to the table}

\item{title}{If \code{TRUE}, add a title to the table header. Default to the same
value than header, except outside of a chunk where it is \code{FALSE} if a table
caption is detected (\code{tbl-cap} YAML entry).}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{show.signif.stars}{If \code{TRUE}, add the significance stars
to the table. The default value is obtained from
\verb{getOption("show.signif.stars"")}.}

\item{...}{Additional arguments (unused for now).}

\item{env}{The environment where to evaluate lazyeval expressions (unused for
now).}
}
\value{
A \strong{flextable} object you can print in different forms or rearrange
with the {flextable} functions.
}
\description{
\code{\link[=tabularise]{tabularise()}} an \strong{htest} object (intoa a \strong{flextable}) that can be
further post-edited..
}
\examples{
data(iris)
iris_cor <- cor.test(iris$Sepal.Length, iris$Sepal.Width)
tabularise::tabularise(iris_cor)

tabularise::tabularise(t.test(x = 1:10, y = 7:20), lang = "fr")
}
